// dente/src/dente-api.d.ts
export type Patient = {
  id: number
  full_name: string
  dob: string
  gender: string
  address: string | null
  phone: string | null
  email: string | null
  created_at: string
  updated_at: string
}

export type PatientCreateUpdate = Omit<Patient, 'id' | 'created_at' | 'updated_at'> & {
  address?: string | null
  phone?: string | null
  email?: string | null
}

export type Procedure = {
  id: number
  patient_id: number
  procedure_name: string
  procedure_date: string
  cost: number
  paid: number
  balance: number
  created_at: string
  updated_at: string
}

export type ProcedureCreateUpdate = Omit<Procedure, 'id' | 'patient_id' | 'created_at' | 'updated_at'>